ref: Move order basket item according to hierarchy
